ABM is working but progress bar is not installed

+++ include/ABM/tools.h
@@ -8,10 +8,10 @@
 
 namespace tools{
 
-	float random(float min, float max); //!< A random value in the given range
+	double random(double min, double max); //!< A random value in the given range
 
 }
-inline float tools::random(float min, float max){
+inline double tools::random(double min, double max){
 	    std::random_device rd;
 	    std::mt19937 gen(rd());
 	    std::uniform_real_distribution<> dis(min,max);
